attempt 1 to fix pylint errors by creating file of common test functions
